Old demos cleanup.
