+++ conditions/src/main/scala/explore/conditions/ConditionsPanel.scala
@@ -5,16 +5,7 @@ package explore.conditions
 
 import japgolly.scalajs.react._
 import japgolly.scalajs.react.vdom.html_<^._
-import react.semanticui.elements.button._
-import react.semanticui.colors._
 import react.common._
-import crystal.react.io.implicits._
-import explore.model._
-import scala.scalajs.js
-import js.JSConverters._
-import js.UndefOr._
-import cats.effect._
-import explore.model.AppStateIO._
 import gem.Observation
 
 /*
@@ -72,34 +63,14 @@ final case class ConditionsPanel(
 object ConditionsPanel {
   type Props = ConditionsPanel
 
-  private def renderButton(forTarget: Target, selected: Option[Target]) = {
-    val color = selected.filter(_ == forTarget).map(_ => Blue).orUndefined
-    Button(onClick = AppState.views.target.set(Some(forTarget)), color = color)(forTarget.toString)
-  }
 
-  private def retrievePersons(): IO[Unit] =
-    for {
-      persons <- AppState.actions.persons.query()
-      _ = println(persons)
-      _ <- AppState.views.persons.set(persons)
-    } yield ()
 
   protected val component =
     ScalaComponent
       .builder[ConditionsPanel]("ConditionsPanel")
       .render { _ =>
         <.div(
-          <.div(
-            Button("IQ"),
-            Button(color = Blue)("Button", "Btn"),
-            Button("Button", "Dec")
-          ),
-          AppState.views.target.streamRender(selected =>
-            <.div(
-              List(Target.M81, Target.M51).toTagMod(target => renderButton(target, selected))
-            )
-          ),
-          Button(onClick = retrievePersons())("GraphQL Test")
+
         )
       }
       .build
