+++ explore/src/main/scala/explore/HomeComponent.scala
@@ -7,16 +7,10 @@ import explore.conditions.ConditionsPanel
 import japgolly.scalajs.react._
 import japgolly.scalajs.react.raw.JsNumber
 import japgolly.scalajs.react.vdom.html_<^._
-import react.semanticui.elements.button._
 import react.common._
 import react.gridlayout._
 import react.sizeme._
 import model._
-import explore.todo.ToDos
-import explore.polls.Polls
-import explore.model.AppStateIO._
-import crystal.react.StreamRenderer
-import crystal.react.io.implicits._
 import gem.Observation
 import gem.ProgramId
 import gsp.math.Index
@@ -24,17 +18,17 @@ import gsp.math.Index
 object HomeComponent {
   private val layoutLg: Layout = Layout(
     List(
-      LayoutItem(x = 0, y = 0, w  = 6, h  = 9, i = "tpe"),
-      LayoutItem(x = 6, y = 0, w  = 6, h  = 9, i = "coords"),
+      LayoutItem(x = 0, y = 0, w  = 6, h  = 10, i = "tpe"),
+      LayoutItem(x = 6, y = 0, w  = 6, h  = 10, i = "coords"),
       LayoutItem(x = 0, y = 10, w = 12, h = 8, i = "doc", isDraggable = false)
     )
   )
 
   private val layoutMd: Layout = Layout(
     List(
-      LayoutItem(x = 0, y = 0, w = 5, h  = 5, i = "tpe"),
-      LayoutItem(x = 6, y = 0, w = 5, h  = 5, i = "coords"),
-      LayoutItem(x = 0, y = 6, w = 10, h = 6, i = "doc", isDraggable = false)
+      LayoutItem(x = 0, y = 0, w = 5, h  = 8, i = "tpe"),
+      LayoutItem(x = 6, y = 0, w = 5, h  = 8, i = "coords"),
+      LayoutItem(x = 0, y = 8, w = 10, h = 6, i = "doc", isDraggable = false)
     )
   )
 
@@ -45,9 +39,6 @@ object HomeComponent {
       // (BreakpointName.sm, (768, 8, layout)),
       // (BreakpointName.xs, (480, 6, layout))
     )
-
-  private val pollConnectionStatus =
-    StreamRenderer.build(AppState.clients.polls.statusStream, Reusability.derive)
 
   val component =
     ScalaComponent
@@ -68,7 +59,7 @@ object HomeComponent {
               layouts          = layouts
             )(
               <.div(^.key := "tpe",    ^.cls := "tile", Tile(Tile.Props("Conditions"), 
-                ConditionsPanel(Observation.Id(ProgramId.Science.fromString.getOption("Program").get, Index.One))
+                ConditionsPanel(Observation.Id(ProgramId.Science.fromString.getOption("GS-2020A-DS-1").get, Index.One))
               )),
               <.div(^.key := "coords", ^.cls := "tile", Tile(Tile.Props("Coordinates"), Imag())),
               <.div(
@@ -76,12 +67,6 @@ object HomeComponent {
                 ^.cls := "tile",
                 Tile(
                   Tile.Props("Target Position"),
-                  <.span(
-                    pollConnectionStatus(status => <.div(s"Poll connection is: $status")),
-                    Button(onClick = AppState.clients.polls.close())("Close Connection")
-                  ),
-                  Polls(p.polls),
-                  ToDos(p.todoList),
                   <.div(p.target.whenDefined(Tpe(_)))
                 )
               )
